Fix logic bug - don't exit all solutions if the current one fails to work - just stop working on that one - it won't work, but siblings might.
Make code clearer by naming code blocks and making them functions.

Use updated Board name GetAdjacentCellIndicies

The biggest change is the additon of NoTrappedCells. This is an important function to cut down on the number of paths visited. By determining that a cell is isolated from any valid flow path, we can ignore the change that brought us here. This is true if there are no paths to/from the cell, and it is true if a cell has a single entry point but will not connect an active node path to its destination.